% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_reactivity.R
\name{\%()\%}
\alias{\%()\%}
\title{Access value of list of reactives}
\usage{
lhs \%()\% rhs
}
\arguments{
\item{lhs}{[list]
list of reactive values}

\item{rhs}{[symbol]
name of the element of the list being the reactive value, unquoted}
}
\value{
A value of reactive expression
}
\description{
Because typing `some_list[["some_values"]]()` takes too long
}
