% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_package.R
\docType{package}
\name{HaDeXGUI-package}
\alias{HaDeXGUI-package}
\alias{HaDeXGUI}
\title{GUI for HaDeX app}
\description{
Under heavy development.
}
\author{
\strong{Maintainer}: Weronika Puchala \email{puchala.weronika@gmail.com} (\href{https://orcid.org/0000-0003-2163-1429}{ORCID})

Authors:
\itemize{
  \item Michal Burdukiewicz \email{michalburdukiewicz@gmail.com} (\href{https://orcid.org/0000-0001-8926-582X}{ORCID})
  \item Dominik Rafacz \email{d.rafacz@student.mini.pw.edu.pl} (\href{https://orcid.org/0000-0003-0925-1909}{ORCID})
}

}
