% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server.R
\name{input_r}
\alias{input_r}
\title{Get input as a reactive value}
\usage{
input_r(name, env = parent.frame())
}
\arguments{
\item{name}{name of the input value}
}
\description{
Get input as a reactive value
}
