% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_server.R
\name{invoke_settings_servers}
\alias{invoke_settings_servers}
\title{Invoke multiple settings servers with automatic parameter binding}
\usage{
invoke_settings_servers(names, modes = character(), env = parent.frame())
}
\arguments{
\item{names}{[character]
a vector of names of servers. This should be only the direct name of a
server, eg. name of `mod_settings_*_server` should be provided as
`"*"`.}

\item{modes}{[named character]
a list of modes passed to corresponding servers. Empty vector by default.
by default.}

\item{env}{[environment]
an environment in which the servers should be installed. Calling
environment by default.}
}
\description{
Invoke multiple settings servers with automatic parameter binding
}
\details{
This function is meant to be called for its side effects. Calling
multiple servers with parameters of standard names is equivalent to calling
this function. This function creates one object named `s_*` per each server
name provided, where `*` is the name of the server. This object is a return
value of the server.
}
