% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_reexport.R
\name{round_any}
\alias{round_any}
\title{Function to round x to nearest multiple of given number}
\usage{
round_any(x, accuracy, f = round)
}
\arguments{
\item{x}{[numeric] a vector of numbers}

\item{accuracy}{[numeric] a number to round to (or a vector of numbers)}

\item{f}{[function] rounding function}
}
\value{
Value of x rounded to nearest multiple of accuracy using function f
}
\description{
Function to round x to nearest multiple of given number
}
\details{
Function reexported from package `{plyr}`
}
